Классы распределены по новым модулям, добавлены модули с тестами для них. Старые удалены. Требуется доработка компиляции
